Refactor and add react-router
