+++ packages/client-gear-playground/package.json
@@ -22,6 +22,7 @@
     "font-awesome": "^4.7.0",
     "react": "15.4.1",
     "react-dom": "15.4.1",
+    "react-router": "3.0.2",
     "webpack": "1.14.0",
     "webpack-dev-server": "1.16.2"
   },
